Decoupling TCP Writer
